+++ docs/source/api/python/random.rst
@@ -1,0 +1,11 @@
+.. _apirandom:
+
+DGL Random Number Generator Controls
+====================================
+
+.. automodule:: dgl.random
+
+.. autosummary::
+    :toctree: ../../generated
+
+    seed
